DICE/requirements.txt
datasets==3.2.0
easydict==1.13
fire==0.6.0
lamb==0.0.1.post1
matplotlib==3.10.0
numpy==2.2.0
Pillow==11.0.0
scikit_learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
torch==2.4.0
torch_npu==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.45.0.dev0
wandb==0.19.1
